Reset listener context after CPZ migration
Currently, during CPZ migration on the DSP, HLOS listener thread
attempts to reconnect. Once CPZ migration is complete, the
adsp_listener_next2 call successfully reaches the DSP. However,
because the context is not reset, it still contains the tidQ of
the caller thread before migration. In the adsp_listener_next2 call
on DSP, prevCtx is checked against the current caller thread’s tidQ.
If not same, an assertion is triggered. However, after CPZ migration,
the current caller tidQ will vary. To avoid this scenario, it is
proposed to reset the context before retrying. If context is set as
zero, check isn't be performed.

Signed-off-by: Bharath Kumar V <quic_bkumar@quicinc.com>