Avoid userspace allocation for add pages request
For add pages request, the capability of kernel to allocate
memory is checked using is_kernel_alloc_supported call. On
newer kernel, this call is expected to always return true.
With changes to use new interface file for QLI, this call
is not returning true which is causing add pages request failure.

Signed-off-by: Bharath Kumar V <quic_bkumar@quicinc.com>