Add fastrpc_test.c, Makefile.am, and README.md for FastRPC test program
This commit introduces the FastRPC test program `fastrpc_test.c` in
the `test` folder. It includes `Makefile.am` to configure the build
process with necessary compiler and linker flags, and a `README.md`
file with instructions for pushing files to the device and running
the test. Additionally, the commit adds `android` and `linux`
directories for shared libraries, and `v68` and `v75` directories
for architecture skeletons. The test program supports options for
domain ID, unsigned PD status, target platform, and architecture
version.

Signed-off-by: Vinayak Katoch <vkatoch@hu-vkatoch-hyd.qualcomm.com>