Avoid issue and PR closure by stale-issues workflow
stale-issues workflow is automatically closing inactive PRs and
issues. Closing of  PRs and issues without any proper resolution
is a bad idea as it would result in loss of valuable context and
also discourage contributors. This also causes missed opportunities
for improvement and increases chances of issue recurrence. Modifying
stale-issues workflow to just mark the inactive PRs and issues as
stale and not close it.

Signed-off-by: Ekansh Gupta <ekansh.gupta@oss.qualcomm.com>